refactor: move weight init from decoders to coca model
